pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hpvd"
version = "1.0.0a1"  # MVP version (1.0.0-alpha1)
description = "Hybrid Probabilistic Vector Database for Trajectory Intelligence"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Kalibry Team"}
]
keywords = ["vector-database", "faiss", "trajectory", "finance", "similarity-search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.4.0",
    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]
ml = [
    "torch>=2.2.0",
    "sentence-transformers>=2.6.0",
    "transformers>=4.41.0",
]
full = [
    "hpvd[dev,ml]",
    "splade>=2.1.0",
    "supabase>=1.0.0",
    "openai>=1.3.0",
]

[project.urls]
Documentation = "https://github.com/kalibry/hpvd-m22/docs"
Repository = "https://github.com/kalibry/hpvd-m22"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

